pyproject.toml
[project]
name = "openkaguya"
version = "0.1.0"
description = "开源版辉夜姬 — 人格化的、拥有主动意识的聊天陪伴助手"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.0",
    "aiohttp>=3.11.0",
    "loguru>=0.7.0",
    "tomli-w>=1.0.0",
    "pydantic>=2.0.0",
    "apscheduler>=3.10.0,<4",
    "aiosqlite>=0.20.0",
    "sqlite-vec>=0.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
browser = [
    "browser-use>=0.11.0",
]

[project.scripts]
kaguya = "kaguya.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kaguya"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
